Add component volType for volume type
ZEN-22701

It would be ok without volume type if we only do normal modeling,
since volume would have a type string for volume type. However,
if OSI admin creates a new volume type, and use the type to create
a new volume, during incremental modeling, we would have
a UUID for the volume type, which is very confusing. Having a
volume type as component fixes this problem.