Add a HostMap class.
Its job is to take in all host identifiers (names, IPs, things with weird prefixes or suffixes)
and generate an authoritative mapping of all of those IDs down to a set of host IDs
as used in the device model.

It should be deterministic, that is, a given input should always produce the same
host ID, unless that value is known to be wrong.

It also generates a single canonical host name for each of these IDs.  This is allowed
to be less deterministic, and thus these names are to be used for labels only.

The modeler's collect() method was was modified to pre-process all the raw
data through the host mapper, and then to deal entirely in IDs in the process() method.