Add ApiEndpoint component type and basic monitoring of the keystone endpoint
ApiEndpoint represents a URL upon which an openstack API can be accessed.
At this time, only the main keystone endpoint is automatically modeled.

The user may specify additional URLs to monitor through a zproperty at this time.
Future versions of the zenpack may discover all the endpoints via the keystone
service catalog.

Monitoring is performed through simple twisted http calls, using
keystone authentication where needed.   Typically, a request
is done to the top level url for the API, and a response is returned
by openstack with a list of available versions.  We consider any
response containing 'version' to be valid.  This is just basic up/down
monitoring of the api endpoint.