Fixes ZPS-1709: A single host can sometimes be modeled as multiple hosts despite them resolving to the same IP
Address situation where host "same as" needed to be transitive- that is,
host A is the same as host B, host B is the same as host C, so
host A, B, and C are all the same.

Add unit test for the exact scenario and add more logging to hostmap
for troubleshooting.