Remove @transact decorator on create_proxy_device to avoid nesting transactions
Nesting @transact will cause these types of low level transaction failures.

This function was originally decorated because proxy_device() was allowed to
create devices as a side effect (that is, a read-only operation could
cause a device to come into existance).  The transact decorator was
used to insure that this gets committed, wherever it might happen.

This was a bad idea for multiple reasons, and so now proxy_device() is
allowed to return a None, and the devices are created only when set_maintain_proxydevices
or set_ensure_service_monitoring get called, and are committed through
normal ApplyDataMap transaction.

Everyplace that calls proxy_device, directly or indirectly, should now
handle the possibility of 'None' being returned.  It wasn't really
that big of a deal.