Feature/zps 1750 (#426)
* Ensure that we don't try to tag the events with a 'None' uuid.

* Attempt to make event transforms more tolerant of malformed data.

I think that this whole module needs a lot of refactoring (it's hard
to follow), but I have tried to make it tolerate a few issues with
incoming data that occur if the user does not isntall our modified
event_definitions.yaml file.  In some cases, the traits are
present, but with different names, so we can still mostly-work
even if the file hasn't been installed.

For instance, in neutron events, the default trait for the ID
is "resource_id", but we called it "id".  So we can just check
both.

For instances, the default one is missing display_name, but we
can use the ID in place of the name.