Prevent modeling bogus host components
These hosts fall into two categories:

* Hosts which only appear in the cinder or nova API URLs

If we don't see a second source of information that the ip/hostname used
In an API url is actually a host (not a load balancer), do not model it by default.

The user can control this by setting zOpenStackNovaApiHosts or zOpenStackCinderApiHosts appropriately.  This confirms to us that the specified
hosts are actual hosts, not a load balancer VIP.

* Ignore the bogus hostname specified in the cinder-volume service if the backend type is "ceph".  (ZPS-3751)

In this case, model the cinder-volume service with no host

Fixes ZPS-3751
Fixes ZPS-3971
Fixes ZPS-4183