update impact relationships for Impact 5.5
Impact 5.5 and DynamicView 1.7 introduced changes that no longer make it
necessary to have specific Impact adapters for component types that
aren't registered to the DynamicView "service_view". So this allows us
to remove all of the impact adapters from this ZenPack, and use
exclusively DynamicView adapters.

DynamicView 1.7 also adds the DynamicViewTestCase base test case that
makes it easier for us to comprehensively test that our defined impact
relationships are correct. So I migrated the test_impact.py tests into
test_dynamicview.py to make use of this.